Move ingest relation tests to different file and test relation extraction
